+++ src/BatchGridModal.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace thetagteam\batchactionsgrid;
-
-class BatchGridModal
-{
-
-}