+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## 0.1
 
 - Initial release
+
+## 0.2
+
+- Remove class BatchGridModal
+- Improve ajax post call without encode ids in json object
+- Add extra data from custom form source